pyproject.toml
[project]
name = "aedist"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0",
    "pyyaml",
    "pypdf",
    "pdf2image",
    "tiktoken",
]

[project.optional-dependencies]
dev = ["pylint", "mypy", "types-PyYAML"]
